move the function implementation to cpp file to avoid multiple definitions
